requirements.txt
aiogram==2.8
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
boilerpy3==1.0.2
certifi==2020.4.5.1
chardet==3.0.4
django-environ==0.4.5
ffmpeg-python==0.2.0
future==0.18.2
idna==2.9
langdetect==1.0.8
multidict==4.7.6
pydub==0.24.0
pytest==6.0.1
python-slugify==4.0.0
pytz==2020.1
six==1.15.0
text-unidecode==1.3
yarl==1.4.2
